Check whether any class of x is dist
class(x) returns a character vector.  When that vector was longer than one,
the original code just checked the first element for a match to "dist" and
issued a warning.  Happens in common case where x is a matrix (and array).
This patch removes the warning for objects with multiple classes.
It also fixes a bug when object's first class is not "dist" but a
subsequent one is.